GrosmoGUI_HealthCareApplication
The main component of our project Grosmo, is a Java-based GUI made by using the Swing module, in which we have made 7 frames, I.e., Start, login, Signup, Register, Patient Details, About, Assistance, etc. Using all these JFrames we are continuously saving and storing the data in MySql(JDBC connection of java) with the help of the XAMPP server and PHP admin. The main idea was to use that data accordingly to show output using Machine Learning.