GrosmoLiveChattingApplication
For assistance and live support to our users and patients as in the health care industry live assistance and live support is very crucial, and to solve the issue we have made the live server for the same where users can chat with our team members on the server to talk, help, issue or anything. It’s the live chat in real-time made using the Socket module in Java.